Add Darts score calculation and test
Implemented a function in Darts.cs that computes the score for a Darts game given the x and y coordinates of the dart's location. Corresponding tests to validate the functionality of the newly added darts score calculation function are added in DartsTests.cs. A README file was also created that includes instructions for the Darts exercise.