Add SpaceAge project and Roman Numerals README
This commit adds a new project, "SpaceAge", to the solution, including necessary package references. It also removes unused namespaces in the 'SpaceAge.cs' file. In addition, a README file has been added to the "Roman Numerals" exercise to provide a detailed instruction for users on how to convert numbers from Arabic to Roman.