Add 'book-store' project with relevant tests and project files
This commit includes the addition of the 'book-store' project to calculate the total cost of books purchased with discounts based on various conditions. This includes the addition of a new csproj file, associated C# code, unit tests to confirm functionality, and a README file to explain the concept. The book-store project has also been added to the main solution file.