Add 'WizardsAndWarriors' game and related components
This commit includes the addition of the 'WizardsAndWarriors' game, which allows you to play as either a Wizard or a Warrior. This includes the introduction of a new csproj file, C# code with classes and methods defining game characters and their behaviors, unit tests to ensure gameplay functionality and a README explaining the game concepts. Additionally, the WizardsAndWarriors project has also been added to the primary solution file.