Add 'Change' project, coin change functionality, and relevant tests
This commit introduces the 'Change' project to the existing solution. The new functionality allows the calculation of the fewest number of coins needed for a target change from a provided set of coins. This solution takes into account edge cases, like negative target or a target smaller than the smallest coin. An extensive set of tests is also added to ensure correctness under various scenarios.