Add 'Grains' project and associated tests
This commit introduces a new 'Grains' project to the solution which involves calculation of grains on a chessboard where the number on each square doubles. Moreover, tests have been added to ensure the functionality works as expected, including handling out of range exceptions. The solution file has been updated to include this new project.