Add Pangram exercise
Add a new exercise to the C# track, called 'Pangram'. This includes creating the README, test file, source code, and project file. This exercise will allow students to check if an English sentence contains every letter in the alphabet at least once. The pangram function will be used to verify the sentence.