Add Gigasecond exercise
This commits adds a new exercise, Gigasecond, to the C# track. The Gigasecond exercise, complete with a README, test suite, and project file, challenges a learner to find the date and time exactly one gigasecond (one thousand million seconds) after a specific moment. The solution has also been placed in the project.