Add Robot Name generator and tests
Implemented a Robot class that creates a robot with a random name in a given format. Also added a series of unit tests in RobotNameTests.cs to ensure the names are unique, follow the format, can be reset, and handle edge cases. Lastly, created a README.md file to provide instructions about the robot name generator.